design: 로딩 레이아웃 개선
