+++ src/app/components/Loading/Loading.module.scss
@@ -1,4 +1,6 @@
 .container {
+  position: absolute;
+  inset: 0;
   height: 100vh;
   background: #000;
   color: #fff;
