+++ src/app/components/Loading/index.tsx
@@ -1,7 +1,27 @@
 import Image from 'next/image';
 import style from './Loading.module.scss';
+import { useEffect } from 'react';
+import { useRouter } from 'next/navigation';
+import { useRecoilValue } from 'recoil';
+import { gender as genderState, nickname as nicknameState } from '@/app/store';
+import usePoint from '@/app/hooks/usePoint';
 
 export default function Loading() {
+  const router = useRouter();
+  const nickname = useRecoilValue(nicknameState);
+  const gender = useRecoilValue(genderState);
+  const { getResultType } = usePoint();
+
+  useEffect(() => {
+    const timer = setTimeout(() => {
+      router.push(`/result/${nickname}/${getResultType()}/${gender}`);
+    }, 4000);
+
+    return () => {
+      clearTimeout(timer);
+    };
+  }, []);
+
   return (
     <div className={style.container}>
       <div className={style.spinner}>
